+++ sdk/keyvault/azure-keyvault-certificates/samples/issuers_async.py
@@ -25,7 +25,7 @@ from azure.core.exceptions import HttpResponseError
 #
 # 2. Get an issuer (get_issuer)
 #
-# 3. List issuers for the key vault (list_issuers)
+# 3. List issuers for the key vault (list_properties_of_issuers)
 #
 # 4. Update an issuer (update_issuer)
 #
@@ -69,7 +69,7 @@ async def run_sample():
         # Now we will list all of the certificate issuers for this key vault. To better demonstrate this, we will first create another issuer.
         await client.create_issuer(name="issuer2", provider="Test", account_id="keyvaultuser", enabled=True)
 
-        issuers = client.list_issuers()
+        issuers = client.list_properties_of_issuers()
 
         async for issuer in issuers:
             print(issuer.name)