+++ sdk/keyvault/azure-keyvault-certificates/tests/test_certificates_client.py
@@ -297,7 +297,7 @@ class CertificateClientTests(KeyVaultTestCase):
                     raise ex
 
         # list certificates
-        result = client.list_certificates()
+        result = client.list_properties_of_certificates()
         self._validate_certificate_list(certificates=result, expected=expected)
 
     @ResourceGroupPreparer(name_prefix=name_prefix)
@@ -335,7 +335,7 @@ class CertificateClientTests(KeyVaultTestCase):
                     raise ex
 
         # list certificate versions
-        self._validate_certificate_list(certificates=(client.list_certificate_versions(cert_name)), expected=expected)
+        self._validate_certificate_list(certificates=(client.list_properties_of_certificate_versions(cert_name)), expected=expected)
 
     @ResourceGroupPreparer(name_prefix=name_prefix)
     @VaultClientPreparer()
@@ -676,7 +676,7 @@ class CertificateClientTests(KeyVaultTestCase):
         )
         expected_issuers = [expected_base_1, expected_base_2]
 
-        issuers = list(client.list_issuers())
+        issuers = list(client.list_properties_of_issuers())
         self.assertEqual(len(issuers), len(expected_issuers))
         for issuer in issuers:
             exp_issuer = next((i for i in expected_issuers if i.name == issuer.name), None)